Crate mu

A wrapper merging the official lambda_runtime and the trustworthy aws_lambda_event crate. The idea behind μ-rs is to provide an easy-to-use api for AWS Serverless Developers, leveraging enterprise-grade semantics in the powerful Rust ecosystem.

The current state of the official AWS Lambda Runtime for Rust provides a good base for developers to runtime simple lambda functions, but when it comes to maintaining a bigger code base, the development experience is not so pleasant. It is especially true if you need to share state between sub-sequent Lambda requests (e.g. Database Connection Pools). μ-rs takes advantage of the official lambda_runtime, providing a consistent API to leverage Lambda development, with especial attention to HTTP Requests.

Modules

alb
An AWS Application Load Balancer abstraction to handle Http requests with AWS Lambda. Being designed with correctness and robustness in mind, it covers more complex scenarios by default - like Http Error handling and customizable serialization mechanism.

lambda
A decorator module that provides consistent API for further Lambda development.