Function mu::alb::runtime::listen_events

pub async fn listen_events<F, Fut, A, B>(handler: F) -> RuntimeResult where
    F: Fn(A) -> Fut + Sync + Send,
    Fut: Future<Output = B> + Send,
    A: Deserialize<A> + Send,
    B: Serialize, 

Listen to ALB events. Unlike mu::lambda::listen_events, this method expects you to respect the AWS Application Load Balancer contract by returning the appropriate response (defined by mu::alb::Response).

It's expected the handler argument to be an async function, as exemplified below.

use mu::{alb, lambda};

#[tokio::main]
async fn main() -> lambda::RuntimeResult {
  alb::listen_events(|req: alb::Request| {
    say_hello()
  }).await
}

async fn say_hello() -> alb::Response {
 alb::response::create_plain_text(
   200, Some("Hello World".to_string()))
}