Function mu::lambda::listen_events

pub async fn listen_events<F, Fut, A, B, E>(handler: F) -> RuntimeResult where
    F: Fn(A, Context) -> Fut + Sync + Send,
    Fut: Future<Output = Result<B, E>> + Send,
    A: for<'de> Deserialize<'de> + Send,
    B: Serialize,
    E: Debug + StdError, 

Listen for AWS Lambda requests of a given A type. This method doesn't enforce any contract aside of the fact it expects a std::result::Result<B, Error> as a return value. It also expects the handler function to be async.

It's important to notice that both A and B types are expected to be Serde deserializable and serializable, respectively.

use aws_lambda_events::event::sqs::SqsEvent;
use mu::lambda;

#[tokio::main]
async fn main() -> lambda::RuntimeResult {
  lambda::listen_events(|sqs_events, ctx| {
    handle_sqs_messages(sqs_events)
  }).await
}

async fn handle_sqs_messages(sqs_events: SqsEvent) -> lambda::LambdaResult<()> {
  println!("Received {} events", sqs_events.records.len());
  Ok(())
}